requirements.txt
flask
flask-socketio
Flask-SQLAlchemy
Pillow
mistune
requests
gunicorn
pymysql
gevent
python-dotenv
